// src/common/enums/postgreSQLErrorsCodes.enum.ts
export enum PostgreSQLErrorCodes {
  UNIQUE_VIOLATION = '23505',
  FOREIGN_KEY_VIOLATION = '23503',
  INVALID_TEXT_REPRESENTATION = '22001',
  INVALID_DATETIME_FORMAT = '22007',
  NOT_NULL_VIOLATION = '23502',
  CHECK_VIOLATION = '23514',
  NUMERIC_VALUE_OUT_OF_RANGE = '22003',
  DIVISION_BY_ZERO = '22012'
}